+++ qutip/qip/device/noise.py
